Fixed compilation warning on 64bit architectures. Thanks to Neal Probert for the hint.
git-svn-id: svn://svn.berlios.de/socketcan/trunk@421 030b6a49-0b11-0410-94ab-b0dab22257f2

+++ kernel/2.6/drivers/net/can/i82527/i82527.c
@@ -1172,7 +1172,8 @@ static __init int i82527_init_module(void)
 
 	if ((sizeof(canmessage_t) != 15) || (sizeof(canregs_t) != 256)) {
 		printk(KERN_WARNING "%s sizes: canmessage_t %d canregs_t %d\n",
-		       CHIP_NAME, sizeof(canmessage_t), sizeof(canregs_t));
+		       CHIP_NAME, (int)sizeof(canmessage_t),
+		       (int)sizeof(canregs_t));
 		return -EBUSY;
 	}
